backend/requirements.txt
flask
flask-cors
PyMuPDF
google-generativeai